+++ Vector.h
@@ -9,7 +9,7 @@
 /* VECTOR HELPER FUNCTIONS */
 void print_vector(std::vector<double> &vec);
 double vector_magnitude(std::vector<double> &vec);
-double vector_dot_product(std::vector<double> &vecA, std::vector<double> &vecB);
+float vector_dot_product(std::vector<double> &vecA, std::vector<double> &vecB);
 std::vector<double> vector_add(std::vector<double> &vecA,
                                std::vector<double> &vecB);
 std::vector<double> vector_sub(std::vector<double> &vecA,
